Dashboard fixes done
